refactor(utils): move formatFactor to @carbon/utils duration helpers
Extracts the factor abbreviation logic from JobTravelerPDF into a
reusable formatFactor(value, unit) util in packages/utils/src/duration.ts,
exported from @carbon/utils alongside the other duration helpers.
